add jwt auth to swagger ui

+++ ressources/item.py
@@ -6,20 +6,30 @@ from db import db
 from sqlalchemy.exc import SQLAlchemyError
 from models import ItemModel
 
+from flask_jwt_extended import jwt_required, get_jwt
+
 blp = Blueprint("items", __name__, description="Operations on items")
 
 
-@blp.route("/item/<string:item_id>")
+@blp.route("/item/<int:item_id>")
 class Item(MethodView):
+    @jwt_required()
     @blp.response(200, ItemSchema)
+    @blp.doc(security=[{"jwt": []}])
     def get(self, item_id):
         """get item"""
         item = ItemModel.query.get_or_404(item_id)
         return item
 
+    @jwt_required()
     @blp.response(200)
+    @blp.doc(security=[{"jwt": []}])
     def delete(self, item_id):
-        """delete item"""
+        """check if user is admin to perform the delete item based on item_id"""
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privilege required.")
+
         item = ItemModel.query.get_or_404(item_id)
         db.session.delete(item)
         db.session.commit()
@@ -27,8 +37,10 @@ class Item(MethodView):
 
         # raise NotImplementedError("Deleteing an item is not implemented.")
 
+    @jwt_required()
     @blp.arguments(ItemUpdateSchema)
     @blp.response(200, ItemSchema)
+    @blp.doc(security=[{"jwt": []}])
     def put(self, item_data, item_id):
         """update item"""
         item = ItemModel.query.get(item_id)
@@ -48,13 +60,17 @@ class Item(MethodView):
 
 @blp.route("/item")
 class ItemList(MethodView):
+    @jwt_required()
     @blp.response(200, ItemSchema(many=True))
+    @blp.doc(security=[{"jwt": []}])
     def get(self):
         """Get all items in db"""
         return ItemModel.query.all()
 
+    @jwt_required()
     @blp.arguments(ItemSchema)
     @blp.response(201, ItemSchema)
+    @blp.doc(security=[{"jwt": []}])
     def post(self, item_data):
         """create an item"""
         new_item = ItemModel(**item_data)
